agents/westfield/requirements.txt
# Google ADK and Gemini
google-genai>=0.1.0

# Reddit API
praw>=7.7.0
prawcore>=2.4.0

# Vector storage and RAG
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Async support
asyncio>=3.4.3
aiohttp>=3.9.0

# CLI and display
click>=8.1.0
rich>=13.0.0
streamlit>=1.29.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Logging
loguru>=0.7.0